refactor top level propCard component
