+++ src/components/proposalActions.jsx
@@ -33,6 +33,7 @@ import { memberVote, MINION_TYPES } from '../utils/proposalUtils';
 import { getTerm, getTitle } from '../utils/metadata';
 import { capitalize, daoConnectedAndSameChain } from '../utils/general';
 import { supportedChains } from '../utils/chain';
+import { earlyExecuteMinionType } from '../utils/minionUtils';
 
 const MotionBox = motion(Box);
 
@@ -66,13 +67,6 @@ const ProposalVote = ({
   const [loading, setLoading] = useState(false);
   const [nextProposalToProcess, setNextProposal] = useState(null);
   const [quorumNeeded, setQuorumNeeded] = useState(null);
-
-  const earlyExecuteMinionType = proposal => {
-    return (
-      proposal?.minion?.minionType === MINION_TYPES.NIFTY ||
-      proposal?.minion?.minionType === MINION_TYPES.SAFE
-    );
-  };
 
   const currentlyVoting = proposal => {
     return (
