+++ src/proposalBuilder/propCardText.js
@@ -2,5 +2,5 @@ export const propStatusText = {
   Unsponsored:
     'A member of the DAO can sponsor this proposal, moving it into the Voting Queue.',
   Cancelled:
-    'This proposal has been struck down, and now it lay beneath the earth, putrifying as the insects begin their feast.',
+    'This proposal has been struck down, and now it lay beneath the earth, putrifying as the insects feast.',
 };
